Refactor CVEList to use a more customizable design pattern
